+++ lecture-2/example3.py
